refactor: split orders lib hosts and ratchet 250-line gate
Give mixed suggestion, validation, and application libraries 250-line
headroom, route leftover ownership scans through the existing SoT helpers,
and pin repo.orders_lib_source_file_size at 250.
