+++ SPEC/program/repo-lint.md
@@ -177,7 +177,7 @@
 | `tool/check_orders_test_support_layout.dart` | Non-`*_test.dart` Dart under `packages/colonizethis_orders/test/orders/` must live under `test/orders/support/` (wave-3 support consolidation; Refs #3949); rule `repo.orders_test_support_layout` |
 | `tool/check_orders_test_support_loc.dart` | Physical-line ceilings for `packages/colonizethis_orders/test/orders/support/**/*.dart` (wave-5 support AC ≤13,950) and package `test/**/*.dart` (wave-9 package AC ≤15,800 after counsel ranking densify; Refs #4109, #4246, #4262, #4281, #4282, #4317, #4307, #4508); rule `repo.orders_test_support_loc` |
 | `tool/check_orders_test_shorthand_file_size.dart` | Physical-line ceiling (**180**) for every `*expectation_shorthand.dart` under `packages/colonizethis_orders/test/orders/support/**` (Refs #4109 wave 5); rule `repo.orders_test_shorthand_file_size` |
-| `tool/check_orders_lib_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_orders/lib/**` physical lines per file (wave-9 ≤300; empty shrink-only grandfather; generated suffixes excluded); rule `repo.orders_lib_source_file_size` (Refs #4246, #4317, #4508) |
+| `tool/check_orders_lib_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_orders/lib/**` physical lines per file (wave-10 ≤250; empty shrink-only grandfather; generated suffixes excluded); rule `repo.orders_lib_source_file_size` (Refs #4246, #4317, #4508, #4587) |
 | `tool/check_app_flame_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/flame/**` physical lines per file (wave-15 ≤300; shrink-only grandfather allowlist empty after slice C; generated suffixes excluded); rule `repo.app_flame_lib_physical_file_size` (Refs #4352) |
 | `tool/check_app_features_screens_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/screens/**` physical lines per file (wave-14 ≤400; shrink-only grandfather during transition; generated suffixes excluded); rule `repo.app_features_screens_lib_physical_file_size` (Refs #4305 Slice B) |
 | `tool/check_orders_test_support_raw_game.dart` | Ceiling on `return Game(` / `=> Game(` under `test/orders/support/**` excluding `common/` (wave-4 migration ratchet; Refs #3971); rule `repo.orders_test_support_raw_game` |
@@ -609,8 +609,8 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a `packages/colonizethis_world/test/world_test_support/` Dart file with more than 280 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_test_support_file_size`, then the run fails naming that file and exits `1` (Refs #4515).
 - Given every non-generated Dart file under `packages/colonizethis_turn/lib/**` has ≤300 physical lines and the shrink-only grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.turn_lib_file_size`, then the rule passes without violations and exits `0` (Refs #4113, #4342).
 - Given a non-generated `packages/colonizethis_turn/lib/**` Dart file with more than 300 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.turn_lib_file_size`, then the run fails naming that file and exits `1` (Refs #4113, #4342).
-- Given every non-generated Dart file under `packages/colonizethis_orders/lib/**` has ≤300 physical lines and the shrink-only grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.orders_lib_source_file_size`, then the rule passes without violations and exits `0` (Refs #4246, #4317, #4508).
-- Given a non-generated `packages/colonizethis_orders/lib/**` Dart file with more than 300 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.orders_lib_source_file_size`, then the run fails naming that file and exits `1` (Refs #4246, #4317, #4508).
+- Given every non-generated Dart file under `packages/colonizethis_orders/lib/**` has ≤250 physical lines and the shrink-only grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.orders_lib_source_file_size`, then the rule passes without violations and exits `0` (Refs #4246, #4317, #4508, #4587).
+- Given a non-generated `packages/colonizethis_orders/lib/**` Dart file with more than 250 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.orders_lib_source_file_size`, then the run fails naming that file and exits `1` (Refs #4246, #4317, #4508, #4587).
 - Given package `packages/colonizethis_orders/test/**/*.dart` physical LOC is at or below 15800, when `dart run tool/ct_repo_lint.dart` runs rule `repo.orders_test_support_loc`, then the rule passes without violations and exits `0` (Refs #4508).
 - Given package `packages/colonizethis_orders/test/**/*.dart` physical LOC exceeds 15800, when `dart run tool/ct_repo_lint.dart` runs rule `repo.orders_test_support_loc`, then the run fails reporting the package test LOC and exits `1` (Refs #4508).
 - Given `packages/colonizethis_economy_test_support/lib/**` has no `*_expectations.dart` files, when `dart run tool/ct_repo_lint.dart` runs rule `repo.economy_test_support_no_expectations_modules`, then the rule passes without violations and exits `0` (Refs #4410).
