+++ packages/colonizethis_orders/lib/src/orders/draft_orders_mutations.dart
@@ -1,5 +1,8 @@
 import 'package:colonizethis_models/colonizethis_models.dart';
 
+export 'draft_orders_mutations_naval.dart';
+export 'draft_orders_mutations_trade.dart';
+
 /// Appends one [DiplomaticOrder] for [playerId] to the turn draft.
 /// SPEC/program/orders.md.
 Orders ordersWithAppendedDiplomaticOrder(
@@ -14,86 +17,6 @@ Orders ordersWithAppendedDiplomaticOrder(
     diplomaticOrdersByPlayerId: {
       ...orders.diplomaticOrdersByPlayerId,
       playerId: list,
-    },
-  );
-}
-
-/// Applies a human naval mission to the turn draft: replaces any prior naval
-/// mission for the same [fleetId] and removes naval move orders for that fleet.
-/// SPEC/program/naval-movement-resolution.md; Refs #4213.
-Orders applyNavalMissionOrderForPlayer(
-  Orders orders,
-  String playerId,
-  NavalMissionOrder newOrder,
-) {
-  final nextMissions = List<NavalMissionOrder>.from(
-    orders.navalMissionOrdersByPlayerId[playerId] ?? const [],
-  )..removeWhere((o) => o.fleetId == newOrder.fleetId);
-  nextMissions.add(newOrder);
-
-  final nextMoves = List<NavalMoveOrder>.from(
-    orders.navalMoveOrdersByPlayerId[playerId] ?? const [],
-  )..removeWhere((o) => o.fleetId == newOrder.fleetId);
-
-  return orders.copyWith(
-    navalMissionOrdersByPlayerId: {
-      ...orders.navalMissionOrdersByPlayerId,
-      playerId: nextMissions,
-    },
-    navalMoveOrdersByPlayerId: {
-      ...orders.navalMoveOrdersByPlayerId,
-      playerId: nextMoves,
-    },
-  );
-}
-
-/// Removes a pending naval mission for [fleetId] from the turn draft.
-/// No-op when no mission is staged. Refs #4213.
-Orders removeNavalMissionOrderForPlayer(
-  Orders orders,
-  String playerId,
-  String fleetId,
-) {
-  final prior =
-      orders.navalMissionOrdersByPlayerId[playerId] ?? const <NavalMissionOrder>[];
-  final next = [
-    for (final NavalMissionOrder o in prior)
-      if (o.fleetId != fleetId) o,
-  ];
-  if (next.length == prior.length) return orders;
-  return orders.copyWith(
-    navalMissionOrdersByPlayerId: {
-      ...orders.navalMissionOrdersByPlayerId,
-      playerId: next,
-    },
-  );
-}
-
-/// Applies a human naval move to the turn draft: replaces any prior naval move for
-/// the same [fleetId] and removes naval mission orders for that fleet.
-/// SPEC/program/naval-movement-resolution.md.
-Orders applyNavalMoveOrderForPlayer(
-  Orders orders,
-  String playerId,
-  NavalMoveOrder newOrder,
-) {
-  final nextMoves = List<NavalMoveOrder>.from(
-    orders.navalMoveOrdersByPlayerId[playerId] ?? const [],
-  )..removeWhere((o) => o.fleetId == newOrder.fleetId);
-  nextMoves.add(newOrder);
-
-  final nextMissions = List<NavalMissionOrder>.from(
-    orders.navalMissionOrdersByPlayerId[playerId] ?? const [],
-  )..removeWhere((o) => o.fleetId == newOrder.fleetId);
-
-  return orders.copyWith(
-    navalMoveOrdersByPlayerId: {
-      ...orders.navalMoveOrdersByPlayerId,
-      playerId: nextMoves,
-    },
-    navalMissionOrdersByPlayerId: {
-      ...orders.navalMissionOrdersByPlayerId,
-      playerId: nextMissions,
     },
   );
 }
@@ -116,14 +39,8 @@ Orders applyCivilianMoveOrderForPlayer(
   )..removeWhere((o) => o.unitId == newOrder.unitId);
 
   return orders.copyWith(
-    moveOrdersByPlayerId: {
-      ...orders.moveOrdersByPlayerId,
-      playerId: nextMoves,
-    },
-    workOrdersByPlayerId: {
-      ...orders.workOrdersByPlayerId,
-      playerId: nextWorks,
-    },
+    moveOrdersByPlayerId: {...orders.moveOrdersByPlayerId, playerId: nextMoves},
+    workOrdersByPlayerId: {...orders.workOrdersByPlayerId, playerId: nextWorks},
   );
 }
 
@@ -146,119 +63,16 @@ Orders applyArmyMoveOrderForPlayer(
   );
 }
 
-/// Drops naval mission orders for fleets that have a naval move order this turn.
-/// SPEC/program/naval-movement-resolution.md.
-Map<String, List<NavalMissionOrder>> navalMissionOrdersRespectingNavalMoves(
-  Map<String, List<NavalMissionOrder>> navalMissionOrdersByPlayerId,
-  Map<String, List<NavalMoveOrder>> navalMoveOrdersByPlayerId,
-) {
-  final out = <String, List<NavalMissionOrder>>{};
-  navalMissionOrdersByPlayerId.forEach((playerId, list) {
-    final movedFleetIds = {
-      for (final o in navalMoveOrdersByPlayerId[playerId] ?? const <NavalMoveOrder>[])
-        o.fleetId,
-    };
-    final filtered = [
-      for (final o in list)
-        if (!movedFleetIds.contains(o.fleetId)) o,
-    ];
-    if (filtered.isNotEmpty) out[playerId] = filtered;
-  });
-  return out;
-}
-
 /// Removes the pending civilian work order at [index] for [playerId] in this
 /// turn's draft [orders]. No-op if the list is missing or [index] is out of
 /// range. SPEC/program/orders.md.
-Orders removePendingWorkOrderAt(
-  Orders orders,
-  String playerId,
-  int index,
-) {
+Orders removePendingWorkOrderAt(Orders orders, String playerId, int index) {
   final list = orders.workOrdersByPlayerId[playerId];
   if (list == null || index < 0 || index >= list.length) {
     return orders;
   }
   final next = List<WorkOrder>.from(list)..removeAt(index);
   return orders.copyWith(
-    workOrdersByPlayerId: {
-      ...orders.workOrdersByPlayerId,
-      playerId: next,
-    },
-  );
-}
-
-/// Returns the pending [TradeOrder] keyed by [commodityId] for [playerId]
-/// in the turn draft [orders], or `null` if none is staged.
-///
-/// SPEC/game/world-market.md § Trade orders. The Trade Screen Market tab
-/// (#2993 E5b) keys staged trade orders by commodity id (one TradeOrder
-/// per (player, commodityId)); this helper exposes that lookup so the
-/// UI can render the current direction (bid / offer) and quantity for
-/// a row without scanning the full list.
-TradeOrder? tradeOrderForPlayerCommodity(
-  Orders orders,
-  String playerId,
-  CommodityId commodityId,
-) {
-  final list = orders.tradeOrdersByPlayerId[playerId];
-  if (list == null) return null;
-  for (final order in list) {
-    if (order.commodityId == commodityId) return order;
-  }
-  return null;
-}
-
-/// Adds or replaces a single [TradeOrder] for [playerId] keyed by
-/// `order.commodityId` in the turn draft [orders].
-///
-/// **Mutual exclusion** (SPEC/game/world-market.md § Trade orders):
-/// any prior pending [TradeOrder] for the same commodity (regardless of
-/// `bid` vs `offer`) is removed before [order] is appended, so each
-/// (player, commodityId) pair carries at most one staged direction. This
-/// matches the per-commodity mutual-exclusion rule enforced later by
-/// [TradeOrderValidator] (#2989) — staging the orders this way means the
-/// UI can never produce a draft that the validator would reject under
-/// rule 3 (`trade_order_mutual_exclusion`).
-Orders applyTradeOrderForPlayer({
-  required Orders orders,
-  required String playerId,
-  required TradeOrder order,
-}) {
-  final priorList = orders.tradeOrdersByPlayerId[playerId] ??
-      const <TradeOrder>[];
-  final next = <TradeOrder>[
-    for (final TradeOrder o in priorList)
-      if (o.commodityId != order.commodityId) o,
-    order,
-  ];
-  return orders.copyWith(
-    tradeOrdersByPlayerId: {
-      ...orders.tradeOrdersByPlayerId,
-      playerId: next,
-    },
-  );
-}
-
-/// Removes any pending [TradeOrder] for [playerId] keyed by
-/// [commodityId] in the turn draft [orders]. No-op when no matching
-/// order is staged. SPEC/game/world-market.md § Trade orders.
-Orders removeTradeOrderForPlayer({
-  required Orders orders,
-  required String playerId,
-  required CommodityId commodityId,
-}) {
-  final priorList = orders.tradeOrdersByPlayerId[playerId];
-  if (priorList == null || priorList.isEmpty) return orders;
-  final filtered = <TradeOrder>[
-    for (final TradeOrder o in priorList)
-      if (o.commodityId != commodityId) o,
-  ];
-  if (filtered.length == priorList.length) return orders;
-  return orders.copyWith(
-    tradeOrdersByPlayerId: {
-      ...orders.tradeOrdersByPlayerId,
-      playerId: filtered,
-    },
+    workOrdersByPlayerId: {...orders.workOrdersByPlayerId, playerId: next},
   );
 }
