+++ packages/colonizethis_orders/lib/src/orders/order_suggestion_research.dart
@@ -3,6 +3,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 
 import 'package:colonizethis_world/colonizethis_world.dart';
 import 'order_suggestion_context.dart';
+import 'order_suggestion_research_diversify.dart';
 
 /// Default research slot count when [Player.researchSlots] is null.
 /// SPEC/game/research-state.md (default 3, 4 with University).
@@ -114,19 +115,21 @@ List<ResearchOrder> suggestResearchOrders(
   // Full-AI category diversification (Refs #3472): off when weight <= 0, which
   // is the default for human / simple-AI / tooling callers (pure greedy).
   final diversify = categoryDiversifyWeight > 0;
-  final bucketWeights = <_AiBucket, int>{
-    _AiBucket.naval: researchNavalWeight,
-    _AiBucket.military: researchMilitaryWeight,
-    _AiBucket.economic: researchEconomicWeight,
-    _AiBucket.exploration: researchExplorationWeight,
+  final bucketWeights = <ResearchAiBucket, int>{
+    ResearchAiBucket.naval: researchNavalWeight,
+    ResearchAiBucket.military: researchMilitaryWeight,
+    ResearchAiBucket.economic: researchEconomicWeight,
+    ResearchAiBucket.exploration: researchExplorationWeight,
   };
 
   // Buckets already represented by lower slots (pending orders + assignments
   // made earlier in this pass) so slots >= 1 can prefer an unrepresented one.
-  final represented = <_AiBucket>{};
+  final represented = <ResearchAiBucket>{};
   for (final o in existingForPlayer) {
     final tech = techCatalog[o.techId];
-    if (tech != null) represented.add(_bucketForCategory(tech.category));
+    if (tech != null) {
+      represented.add(researchBucketForCategory(tech.category));
+    }
   }
 
   var inProgressCursor = 0;
@@ -135,7 +138,7 @@ List<ResearchOrder> suggestResearchOrders(
     if (inProgressCursor < inProgressTechIds.length) {
       techId = inProgressTechIds[inProgressCursor++];
     } else if (pool.isNotEmpty) {
-      techId = _pickNewTech(
+      techId = pickDiversifiedResearchTech(
         slotIndex: slot,
         pool: pool,
         diversify: diversify,
@@ -149,7 +152,9 @@ List<ResearchOrder> suggestResearchOrders(
       break;
     }
     final tech = techCatalog[techId];
-    if (tech != null) represented.add(_bucketForCategory(tech.category));
+    if (tech != null) {
+      represented.add(researchBucketForCategory(tech.category));
+    }
     suggestions.add(
       ResearchOrder(
         slotIndex: slot,
@@ -164,92 +169,4 @@ List<ResearchOrder> suggestResearchOrders(
     'inProgress=${inProgressTechIds.length} suggested=${suggestions.length}',
   );
   return suggestions;
-}
-
-/// AI category buckets that group the seven game tech categories for Full-AI
-/// research diversification. SPEC/program/order-suggestions.md § Research orders.
-enum _AiBucket { naval, military, economic, exploration }
-
-/// Fixed tiebreak order for equal-weight buckets: naval > military > economic >
-/// exploration. Also the deterministic iteration order for bucket selection.
-const List<_AiBucket> _bucketOrder = <_AiBucket>[
-  _AiBucket.naval,
-  _AiBucket.military,
-  _AiBucket.economic,
-  _AiBucket.exploration,
-];
-
-/// Maps a game tech category to its AI bucket. Unmapped categories (and
-/// `new-world` / `diplomatic` / `civilian`) fall back to exploration.
-_AiBucket _bucketForCategory(String category) {
-  switch (category) {
-    case 'naval':
-    case 'transport':
-      return _AiBucket.naval;
-    case 'military':
-      return _AiBucket.military;
-    case 'gathering':
-    case 'labour':
-      return _AiBucket.economic;
-    default:
-      return _AiBucket.exploration;
-  }
-}
-
-/// Deterministic 0..99 roll from [seed] (stable pure-int hash; no RNG state).
-int _diversifyRoll(int seed) {
-  var x = seed & 0x7fffffff;
-  x = (x * 1103515245 + 12345) & 0x7fffffff;
-  return x % 100;
-}
-
-/// Greedy-cheapest tech id for [pool]'s head, or — for slots `>= 1` when
-/// [diversify] and the per-slot blend rolls in — the greedy-first tech in the
-/// highest-weight AI bucket not yet [represented]. Falls back to greedy when
-/// the chosen bucket has no available tech. [pool] is greedy-sorted by caller.
-String _pickNewTech({
-  required int slotIndex,
-  required List<TechDefinition> pool,
-  required bool diversify,
-  required Map<_AiBucket, int> bucketWeights,
-  required Set<_AiBucket> represented,
-  required int researchSeed,
-  required int diversifyWeight,
-}) {
-  if (!diversify || slotIndex == 0) return pool.first.id;
-  if (_diversifyRoll(researchSeed + slotIndex) >= diversifyWeight) {
-    return pool.first.id;
-  }
-  final bucket = _chooseDiversifiedBucket(
-    pool: pool,
-    bucketWeights: bucketWeights,
-    represented: represented,
-  );
-  if (bucket == null) return pool.first.id;
-  for (final tech in pool) {
-    if (_bucketForCategory(tech.category) == bucket) return tech.id;
-  }
-  return pool.first.id;
-}
-
-/// Highest-weight AI bucket that is both unrepresented and has at least one
-/// tech in [pool], ties broken by [_bucketOrder]. Null when none qualifies.
-_AiBucket? _chooseDiversifiedBucket({
-  required List<TechDefinition> pool,
-  required Map<_AiBucket, int> bucketWeights,
-  required Set<_AiBucket> represented,
-}) {
-  final available = <_AiBucket>{
-    for (final tech in pool) _bucketForCategory(tech.category),
-  };
-  final candidates = _bucketOrder
-      .where((b) => !represented.contains(b) && available.contains(b))
-      .toList();
-  if (candidates.isEmpty) return null;
-  candidates.sort((a, b) {
-    final byWeight = (bucketWeights[b] ?? 0).compareTo(bucketWeights[a] ?? 0);
-    if (byWeight != 0) return byWeight;
-    return _bucketOrder.indexOf(a).compareTo(_bucketOrder.indexOf(b));
-  });
-  return candidates.first;
 }
