+++ tool/ct_repo_lint_manifest.yaml
@@ -1724,7 +1724,7 @@ rules:
 
   - rule_id: repo.orders_lib_source_file_size
     group: structure
-    title: colonizethis_orders lib files must stay at or below the wave-9 300 physical-line ceiling (Refs #4508)
+    title: colonizethis_orders lib files must stay at or below the wave-10 250 physical-line ceiling (Refs #4587)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_orders_lib_source_file_size.dart
